다형성 활용
