+++ src/main/java/flab/delideli/controller/UpdateMemberController.java
@@ -1,14 +1,10 @@
 package flab.delideli.controller;
 
-import flab.delideli.dto.MemberDTO;
 import flab.delideli.dto.UpdateDTO;
-import flab.delideli.service.LoginService;
 import flab.delideli.service.MemberService;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
