+++ src/main/java/flab/delideli/interceptor/LoginInterceptor.java
@@ -1,10 +1,8 @@
 package flab.delideli.interceptor;
 
 import flab.delideli.exception.UnauthorizedException;
-import flab.delideli.service.LoginService;
+import flab.delideli.service.SessionLoginService;
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.ModelAndView;
@@ -16,11 +14,11 @@ import javax.servlet.http.HttpServletResponse;
 @AllArgsConstructor
 public class LoginInterceptor implements HandlerInterceptor {
 
-    private LoginService loginService;
+    private SessionLoginService SessionloginService;
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        String currentUserId = loginService.getSessionUserId(request);
+        String currentUserId = SessionloginService.getSessionUserId(request);
         if (currentUserId != null) {
             return true;
         }
