+++ src/main/java/flab/delideli/service/SessionLoginService.java
@@ -1,0 +1,20 @@
+package flab.delideli.service;
+
+import org.springframework.stereotype.Service;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
+@Service
+public class SessionLoginService implements LoginService{
+
+    private static final String USER_ID = "USER_ID";
+
+    public String getSessionUserId(HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        String currentUserId = (String) session.getAttribute(USER_ID);
+
+        return currentUserId;
+    }
+
+}
